move experiment manipulations to new module manipulation.py
